+++ changedetectionio/content_fetchers/requests.py
@@ -1,7 +1,7 @@
 from loguru import logger
 import hashlib
 import os
-from changedetectionio import strtobool
+from changedetectionio.strtobool import strtobool
 from changedetectionio.content_fetchers.exceptions import BrowserStepsInUnsupportedFetcher, EmptyReply, Non200ErrorCodeReceived
 from changedetectionio.content_fetchers.base import Fetcher
