+++ changedetectionio/tests/test_security.py
@@ -4,7 +4,7 @@ from flask import url_for
 from .util import set_original_response, set_modified_response, live_server_setup, wait_for_all_checks
 import time
 
-from .. import strtobool
+from changedetectionio.strtobool import strtobool
 
 
 def test_setup(client, live_server, measure_memory_usage):
